Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
selenium==4.38.0
undetected-chromedriver==3.5.4
sqlalchemy==2.0.45
asyncpg>=0.27.0
alembic>=1.11.0
setuptools
httpx
pydantic>=2.0
pydantic-settings
sentence-transformers

